refactor urls into folder
