+++ src/lib/dynamicData/badges.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 import type { BadgeDataType } from '@/types/index';
-import { backendUrls } from '@/lib/backendUrls';
+import { backendUrls } from '@/lib/urls';
 
 export function getAllBadges(): Promise<BadgeDataType[]> {
 	return axios
