+++ src/lib/urls/index.ts
@@ -1,0 +1,2 @@
+export { backendUrls } from './backendUrls';
+export { externalUrls } from './externalUrls';
